requirements.txt
Flask>=2.3,<3.0
Flask-Login>=0.6.2
Flask-SQLAlchemy>=3.0.0
Flask-Migrate>=4.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
sib-api-v3-sdk>=7.6.0
email-validator>=2.0.0
gunicorn>=21.0.0
vonage>=3.0.0
